Split history.rs into smaller files
No functional change.
