+++ src/routes/stats/history/filters/mod.rs
@@ -1,0 +1,26 @@
+// Pi-hole: A black hole for Internet advertisements
+// (c) 2018 Pi-hole, LLC (https://pi-hole.net)
+// Network-wide ad blocking via your own hardware.
+//
+// API
+// History Filters
+//
+// This file is copyright under the latest version of the EUPL.
+// Please see LICENSE file for your rights under this license.
+
+mod blocked;
+mod client;
+mod dnssec;
+mod domain;
+mod private;
+mod query_type;
+mod reply;
+mod setup_vars;
+mod status;
+mod time;
+mod upstream;
+
+pub use self::{
+    blocked::*, client::*, dnssec::*, domain::*, private::*, query_type::*, reply::*,
+    setup_vars::*, status::*, time::*, upstream::*
+};
