+++ src/routes/stats/history/mod.rs
@@ -1,0 +1,20 @@
+// Pi-hole: A black hole for Internet advertisements
+// (c) 2018 Pi-hole, LLC (https://pi-hole.net)
+// Network-wide ad blocking via your own hardware.
+//
+// API
+// History Endpoints
+//
+// This file is copyright under the latest version of the EUPL.
+// Please see LICENSE file for your rights under this license.
+
+mod endpoints;
+mod filters;
+mod get_history;
+mod map_query_to_json;
+mod skip_to_cursor;
+
+#[cfg(test)]
+mod testing;
+
+pub use self::endpoints::*;
